Done with basic PSO
It's a working PSO, it doesn't get the most amazing results, but it works. I think it would reach some great results after maybe a few million iterations, or perhaps more particles.

Aside from that a few data related files were added. I've tested on both c101 and c202 so I added them both (the .txt aren't necessary anymore). And then there's the cXXX_to_csv.py which has the contents of c202 and c101 as arrays so it can parse the data to .csv