Updated the PSO
It's better overall, now that it distributes to more vehicles than just 10. But for some reason it calculates time wrong, and I can't figure out why/how.
Usually has correct times at higher/many penalties, but anything which registers lower than 10'000 is somewhat incorrect. The incorrections are unique for each route, so wth is the problem